+++ Sources/Qualtive/Models/Score.swift
@@ -20,7 +20,8 @@ extension Score {
   /// - `smilies3`: 3 user options displayed as smilies
   /// - `thumbs`: 2 user options displayed as up and down thumbs
   /// - `nps`: 11 user options displayed as a range of numbers starting with 0 and ending on 10
-  public enum Kind: Sendable {
+  /// - `stars5`: 5 user options displayed as stars
+  public enum Kind: Sendable, Equatable {
 
     /// 5 user options displayed as smilies
     case smilies5
@@ -33,5 +34,8 @@ extension Score {
 
     /// 11 user options displayed as a range of numbers starting with 0 and ending on 10
     case nps(leadingText: String, trailingText: String)
+
+    /// 5 user options displayed as stars
+    case stars5
   }
 }